handle reservation and errors

+++ reservation/views.py
@@ -14,6 +14,7 @@ from rest_framework.exceptions import ValidationError
 from .models import match_reservation, online_match_payment
 from match.models import Match
 from match.views import update_match
+from account.models import User
 
 # Create your views here.
 
@@ -24,18 +25,28 @@ def book_match(request, pk):
     renderer_classes = [userrenderer]
     permission_classes = [IsAuthenticated]
     try:
-        match_id = Match.objects.get(pk=pk)
+        #get the instance
+        match = Match.objects.get(pk=pk)
+        user = request.user
     except Match.DoesNotExist:
         return Response({"error": "Match not found"}, status=status.HTTP_404_NOT_FOUND)
-    if match_id.avilable == True:
+    if match.no_tickets ==0:
+      match.avilable = False
+      
+    if match.avilable == True:
 
-        serializer = serializers.bookmatchserializer(data=request.data)
+        serializer= serializers.bookmatchserializer(data = request.data)
+        # price =0.0
 
         if serializer.is_valid():
-            serializer.save(user_id=request.user, match_id=match_id)
-            x = match_id.no_tickets
-            match_id.no_tickets -= serializer.data.get("tickets_reserved")
-            if match_id.no_tickets < 0:
+            price =  match.ticket_price * serializer.validated_data.get("tickets_reserved")
+            
+            
+            serializer.save(user_id = request.user,match_id = match,price = price)
+            x=match.no_tickets
+            y=serializer.data.get("tickets_reserved")
+            temp_x = x-y
+            if  temp_x <0:
                 return Response(
                     {
                         "error": [
@@ -45,16 +56,19 @@ def book_match(request, pk):
                     },
                     status=status.HTTP_400_BAD_REQUEST,
                 )
-            if match_id.no_tickets == 0:
-                match_id.avilable = False
-            match_id.save()
-            # serializer.data.get("price")=  match_id.price * serializer.data.get("tickets_reserved")
-            price = match_id.ticket_price * serializer.data.get("tickets_reserved")
-
-            return Response(
-                [{"msg": f"the price is {price}"}, serializer.data],
-                status=status.HTTP_201_CREATED,
-            )
+            if  temp_x ==0:
+                match.avilable = False
+            
+            if serializer.data.get("pay_method")=="offline":
+                # print("yara")
+                match.no_tickets = temp_x
+                match.save()
+                
+                return Response({"message": [f'the price is {price}',"The match has been booked successfully, check your email for payment details and ticket receipt date",serializer.data]}, status=status.HTTP_200_OK)
+            else :
+                
+                return Response([{'msg':[f'the price is {price}']},request.user.username,match.name], status=status.HTTP_201_CREATED)  # it shouldn't appear in reservation page but must appear in payment page
+           
 
         return Response(serializer.errors, status=status.HTTP_406_NOT_ACCEPTABLE)
     else:
@@ -63,7 +77,7 @@ def book_match(request, pk):
         )
 
 
-@api_view(["POST"])
+@api_view(["DELETE"])
 def cancel_reservation(request, pk):
     renderer_classes = [userrenderer]
     permission_classes = [IsAuthenticated]
@@ -81,6 +95,46 @@ def cancel_reservation(request, pk):
             {"error": "You are not the owner of this reservation"},
             status=status.HTTP_400_BAD_REQUEST,
         )
+
+
+#----------------------------------------------------------------------------------------------------------------
+@api_view(['GET'])
+def list_all_reservations_for_admin (request):
+    renderer_classes = [userrenderer]
+    permission_classes = [IsAuthenticated]
+    
+    if request.user.is_staff:
+        reservations = match_reservation.objects.all()
+        serializer = serializers.listadminserializer(reservations, many=True)
+        # serializer.data.get()
+        return Response(serializer.data, status=status.HTTP_200_OK)
+    else:
+        return Response({"error": "You are not authorized to view this page"}, status=status.HTTP_401_UNAUTHORIZED)
+    
+@api_view (['GET'])
+def  list_all_reservations_for_user (request):
+    renderer_classes = [userrenderer]
+    permission_classes = [IsAuthenticated]
+    reservations =  match_reservation.objects.filter(user_id=request.user)
+    serializer =  serializers.listuserserializer(reservations, many=True)
+    return Response(serializer.data, status=status.HTTP_200_OK)
+#----------------------------------------------------------------------------------------------------------------
+@api_view(['PATCH'])
+def update_reservation (request,pk):
+    renderer_classes = [userrenderer]
+    permission_classes = [IsAuthenticated]
+    try:
+        reservation = match_reservation.objects.get(pk=pk)
+    except match_reservation.DoesNotExist:
+        return Response({"error": "Reservation not found"}, status=status.HTTP_404_NOT_FOUND)
+    if reservation.user_id == request.user:
+        serializer = serializers.updatereservationserializer(reservation, data=request.data,partial = True)
+        if serializer.is_valid():
+            serializer.save()
+            return Response({"msg":['updated successfully',serializer.data]}, status=status.HTTP_200_OK)
+        else:
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+    
 
 
 # @api_view(["POST"])
@@ -120,15 +174,17 @@ def match_payment(request, pk):
             {"error": "You already paid for this match"},
             status=status.HTTP_404_NOT_FOUND,
         )
+    x=match.no_tickets
 
     serializer = serializers.matchpaymentserializer(data=request.data)
     if serializer.is_valid():
-        serializer.save(
-            reservation_id=reservation_id
-        )  # Saving the reservation in the payment
-        reservation_id.delete()
-        return Response({"msg": "Payment successful"}, status=status.HTTP_200_OK)
-    else:
-        match.no_tickets += reservation_id.tickets_reserved
+        serializer.save(reservation_id=reservation_id  )  # Saving the reservation in the payment
+        match.no_tickets -= reservation_id.tickets_reserved
+        if  match.no_tickets <0:
+            return Response({"error": ["Not enough tickets",f'you can book up to  {x} tickets only!']}, status=status.HTTP_400_BAD_REQUEST)
+        if match.no_tickets == 0:
+            match.avilable = False
         match.save()
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        reservation_id.delete()
+        return Response({"msg": ["Payment done!",serializer.data]}, status=status.HTTP_200_OK)
+    return Response(serializer.errors, status=status.HTTP_402_PAYMENT_REQUIRED)
